refactoring, not a true class yet but it's getting less awful
